refactor: Move isPair, isList to pair.js
To reduce circular dependency
