+++ src/library/srfi.js
@@ -5,11 +5,10 @@ import { define_libfunc, alias_libfunc, define_syntax, define_scmfunc,
          assert_char, assert_symbol, assert_port, assert_pair, assert_list,
          assert_vector,
          assert_function, assert_closure, assert_procedure, assert_date, assert, deprecate } from "./infra.js"; 
-import { isPair } from "../system/_types.js"
 import { to_write, to_display } from "../system/_writer.js"
 import { write_ss } from "../system/write_ss.js"
 import Console from "../system/console.js"
-import { Pair, array_to_list, deep_array_to_list } from "../system/pair.js"
+import { Pair, isPair, array_to_list, deep_array_to_list } from "../system/pair.js"
 import { Port } from "../system/port.js"
 import { Sym } from "../system/symbol.js"
