+++ src/main.js
@@ -1,7 +1,7 @@
 import { TopEnv, CoreEnv, nil, undef, max_trace_size, suppress_deprecation_warning } from "./header.js";
 import { VERSION, GitCommit } from "./version.js";
 
-import { isNil, isUndef, isBoolean, isString, isFunction, isChar, isSymbol, isPort, isPair, isList,
+import { isNil, isUndef, isBoolean, isString, isFunction, isChar, isSymbol, isPort,
          isVector, isHashtable, isMutableHashtable, isClosure, makeClosure, isProcedure,
          isSelfEvaluating, eq, eqv, equal, lt } from "./system/_types.js"
 import { to_write, to_display, inspect } from "./system/_writer.js"
@@ -14,7 +14,7 @@ import { BiwaError, Bug, UserError } from "./system/error.js"
 import Hashtable from "./system/hashtable.js"
 import Interpreter from "./system/interpreter.js"
 import { Complex, Rational, isNumber, isComplex, isReal, isRational, isInteger } from "./system/number.js"
-import { Pair, List, array_to_list, deep_array_to_list, Cons } from "./system/pair.js"
+import { Pair, List, isPair, isList, array_to_list, deep_array_to_list, Cons } from "./system/pair.js"
 import Parser from "./system/parser.js"
 import Pause from "./system/pause.js"
 import { Port, eof } from "./system/port.js"
